Further small refactoring/cleanup 2
