+++ tests/system/src/test_env.cxx
@@ -19,7 +19,7 @@
 
 #include "build_cxx/system_tests/test_env.hxx"
 
-#include "build_cxx/system_tests/env.hxx"
+#include "build_cxx/system_tests/impl/env.hxx"
 
 namespace build_cxx::system_tests {
